split out ElevateChecker
