+++ src/NServiceBus.WindowsPerformanceCounters/NServiceBus.WindowsPerformanceCounters.csproj
@@ -49,6 +49,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AssemblyInfo.cs" />
+    <Compile Include="Counters\ElevateChecker.cs" />
     <Compile Include="Counters\IPerformanceCounterInstance.cs" />
     <Compile Include="Counters\NonFunctionalPerformanceCounterInstance.cs" />
     <Compile Include="Counters\PerformanceCounterHelper.cs" />
